techwise-test/requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
scikit-learn
pandas
numpy
matplotlib
seaborn
networkx
psycopg2-binary
plotly
tqdm
